Hide Reply All and Conversation Deletion

+++ libs/pandautils/src/main/java/com/instructure/pandautils/features/inbox/details/composables/InboxDetailsScreen.kt
@@ -125,7 +125,7 @@ private fun AppBar(
             }
         } else null,
         actions = {
-            AppBarMenu(uiState.conversation, actionHandler)
+            AppBarMenu(uiState, actionHandler)
         },
         backgroundColor = Color(color = ThemePrefs.primaryColor),
         contentColor = Color(color = ThemePrefs.primaryTextColor),
@@ -290,8 +290,9 @@ private fun InboxDetailsContentView(
 }
 
 @Composable
-private fun AppBarMenu(conversation: Conversation?, actionHandler: (InboxDetailsAction) -> Unit) {
+private fun AppBarMenu(uiState: InboxDetailsUiState, actionHandler: (InboxDetailsAction) -> Unit) {
     var showMenu by rememberSaveable { mutableStateOf(false) }
+    val conversation = uiState.conversation
     OverflowMenu(
         modifier = Modifier
             .background(color = colorResource(id = R.color.backgroundLightestElevated))
@@ -312,13 +313,15 @@ private fun AppBarMenu(conversation: Conversation?, actionHandler: (InboxDetails
                     MessageMenuItem(R.drawable.ic_reply, stringResource(id = R.string.reply))
                 }
 
-                DropdownMenuItem(
-                    onClick = {
-                        showMenu = !showMenu
-                        actionHandler(InboxDetailsAction.ReplyAll(message))
+                if (uiState.showReplyAllButton) {
+                    DropdownMenuItem(
+                        onClick = {
+                            showMenu = !showMenu
+                            actionHandler(InboxDetailsAction.ReplyAll(message))
+                        }
+                    ) {
+                        MessageMenuItem(R.drawable.ic_reply_all, stringResource(id = R.string.replyAll))
                     }
-                ) {
-                    MessageMenuItem(R.drawable.ic_reply_all, stringResource(id = R.string.replyAll))
                 }
             }
 
@@ -402,13 +405,15 @@ private fun AppBarMenu(conversation: Conversation?, actionHandler: (InboxDetails
                 }
             }
 
-            DropdownMenuItem(
-                onClick = {
-                    showMenu = !showMenu
-                    actionHandler(InboxDetailsAction.DeleteConversation(conversation.id))
+            if (uiState.showDeleteButton) {
+                DropdownMenuItem(
+                    onClick = {
+                        showMenu = !showMenu
+                        actionHandler(InboxDetailsAction.DeleteConversation(conversation.id))
+                    }
+                ) {
+                    MessageMenuItem(R.drawable.ic_trash, stringResource(id = R.string.delete))
                 }
-            ) {
-                MessageMenuItem(R.drawable.ic_trash, stringResource(id = R.string.delete))
             }
         }
 
@@ -500,6 +505,7 @@ fun InboxDetailsScreenContentPreview() {
             author = author,
             recipients = recipients,
             enabledActions = true,
+            canDelete = true,
         )
     }
